feat: add button to decrement items in a cart
• The cart is currently hidden when the user empties it by redrawing the
entire catalog. Can you change the application to remove it using an Ajax
request, so the page doesn’t reload?
• Add a button next to each item in the cart. When clicked, it should invoke
an action to decrement the quantity of the item, deleting it from the cart
when the quantity reaches zero. Get it working without using Ajax first
and then add the Ajax goodness.
• Make images clickable. In response to a click, add the associated product
to the cart.
• When a product changes, highlight the product that changed in response
to receiving a broadcast message.